Use size_t for string_{,l,r}trim()
These functions do not fail, and may work on very long strings.
Reflect this by returning unsigned values (number of characters
trimmed).

XXX breaks the ABI

+++ include/System/string.h
@@ -62,8 +62,8 @@ ssize_t string_index(String const * string, String const * key);
 
 int string_replace(String ** string, String const * what, String const * by);
 
-int string_ltrim(String * string, String const * which);
-int string_rtrim(String * string, String const * which);
-int string_trim(String * string, String const * which);
+size_t string_ltrim(String * string, String const * which);
+size_t string_rtrim(String * string, String const * which);
+size_t string_trim(String * string, String const * which);
 
 #endif /* !LIBSYSTEM_SYSTEM_STRING_H */
